Rename rasterIndex -> index
start counting from 1
